+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Playwright tests for navigation menu, what is public domain and search bar, explore further for the home page (DR-3658)
 - Added Playwright test to search for a keyword (DR-3659)
 - Added Playwright test to confirm visibility of search result filters (DR-3659)
+- Added GitHub Actions workflow to run Playwright tests on PRs (DR-3721)
 
 ## [0.4.2] 2025-05-22