+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Playwright test to confirm visibility of search result filters (DR-3659)
 - Added Playwright test to filter search results with dropdowns (DR-3659)
 - Added Image ID to Items page (DR-3695)
+- Added GitHub Actions workflow to run Playwright tests on PRs (DR-3721)
 
 ### Fixed
 - Fixed Order Print button (DR-3486)