+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.4.3] 2025-06-17
 ### Added
 - Added Universal Viewer support and metadata fields
 - Conditionally render viewer with rights logic (DR-3645)